Split analysis resources and tests
