+++ bemserver_api/resources/analysis/__init__.py
@@ -1,4 +1,17 @@
-from .routes import blp  # noqa
+"""Analysis resources"""
+from bemserver_api import Blueprint
+
+from .completeness.routes import blp as completeness_blp
+from .energy_consumption.routes import blp as energy_consumption_blp
+
+
+blp = Blueprint(
+    "Analysis", __name__, url_prefix="/analysis", description="Data analysis operations"
+)
+
+
+blp.register_blueprint(completeness_blp)
+blp.register_blueprint(energy_consumption_blp)
 
 
 def register_blueprints(api):
